Kernel: Xiaomi kernel changes for XiaoMi 4C/XiaoMi Note Pro/XiaoMi 4S.
The Patch based on QualComm release version:LA.BF64.1.2.3-01110-8x94.0

The kernel config file used is libra_user_defconfig/leo_user_defconfig

+++ arch/arm/mm/dma-mapping.c
@@ -1202,7 +1202,7 @@ static struct page **__iommu_alloc_buffer(struct device *dev, size_t size,
 					  gfp_t gfp, struct dma_attrs *attrs)
 {
 	struct page **pages;
-	int count = size >> PAGE_SHIFT;
+	size_t count = size >> PAGE_SHIFT;
 	int array_size = count * sizeof(struct page *);
 	int i = 0;
